move nemo_file into loader folder
