+++ nemo/export/tensorrt_llm.py
@@ -28,7 +28,7 @@ import wrapt
 
 from nemo.deploy import ITritonDeployable
 from nemo.export.tarutils import TarPath, unpack_tarball
-from nemo.export.trt_llm.nemo.nemo_file import get_tokenzier, load_nemo_model
+from nemo.export.trt_llm.loader.nemo_file import get_tokenzier, load_nemo_model
 from nemo.export.trt_llm.nemo_utils import is_nemo_file, model_to_trtllm_ckpt
 from nemo.export.trt_llm.qnemo import qnemo_to_tensorrt_llm
 from nemo.export.trt_llm.qnemo.tokenizer_utils import get_nmt_tokenizer
